Front end work and Recursive opening
The front end is coming along. The game is starting to resemble minesweeper. I got the recursive openings on a 0 working. It is a little slow. Think about the game computing the number of mines surronding each tile at beginning of game to speed up opening mines. That or it will have to be solved another way. (Might be better to solve by not pressing each button, but rather a special function for the recursive operation)